+++ plugins/rpgmaker_mv/__init__.py
@@ -1,0 +1,5 @@
+"""RPG Maker MV plugin package."""
+
+from plugins.rpgmaker_mv.plugin import RPGMakerMVPlugin
+
+__all__ = ["RPGMakerMVPlugin"]